雙散列率

+++ src/main/kotlin/ninja/blacknet/api/StakingInfo.kt
@@ -14,7 +14,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 class StakingInfo(
         val stakingAccounts: Int,
-        val hashRate: Float,
+        val hashRate: Double,
         val weight: String,
         val networkWeight: String,
         val expectedTime: Long
